테이블 재설계후 DTO 수정

+++ backend/CoCo/src/main/java/com/smhrd/coco/domain/TB_NOTICE.java
@@ -14,9 +14,9 @@ import lombok.ToString;
 public class TB_NOTICE {
 
 	
-    private Long NOTICE_ID;  // 알림 번호   
+    private Integer NOTICE_ID;  // 알림 번호   
     private String CUST_ID; // 알림을 받는 회원아이디   
     private String NOTICE_CONTENT; // 알림 내용   
     private String NOTICE_DT; // 알림 전송 일시
-    private Long BOARD_ID;
+    private Integer BOARD_ID; // 게시글 번호 
 }
